Moved the MainJFrame to UI package
