+++ ProfileWorkareaExample/nbproject/project.properties
@@ -74,7 +74,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=ProfileWorkareaExample
-main.class=Business.ProfileWorkAreaMainFrame
+main.class=UserInterface.ProfileWorkAreaMainFrame
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
