+++ ProfileWorkareaExample/src/Business/ConfigureABusiness.java
@@ -21,9 +21,9 @@ import Business.UserAccounts.UserAccountDirectory;
  *
  * @author kal bugrara
  */
-class ConfigureABusiness {
+public class ConfigureABusiness {
 
-    static Business initialize() {
+    public static Business initialize() {
         Business business = new Business("Information Systems");
 
 // Create Persons
