+++ ProfileWorkareaExample/src/UserInterface/ProfileWorkAreaMainFrame.java
@@ -3,8 +3,10 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Business;
+package UserInterface;
 
+import Business.Business;
+import Business.ConfigureABusiness;
 import Business.Profiles.EmployeeProfile;
 import Business.Profiles.Profile;
 import Business.Profiles.StudentProfile;
